requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
anyio=3.6.2=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.5.7=hf0a4a13_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=pyhd8ed1ab_0
charset-normalizer=3.1.0=pypi_0
debugpy=1.5.1=py310hc377ac9_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.12.0=pypi_0
google-auth=2.18.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.2=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=0.16.3=pypi_0
httpx=0.23.3=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
ipykernel=6.15.0=pyh736e0ef_0
ipython=8.13.2=pyhd1c38e8_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310hca03da5_0
libcxx=16.0.4=h4653b0c_0
libffi=3.4.2=hca03da5_6
libsodium=1.0.18=h27ca646_1
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.2=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h03a7124_0
packaging=23.1=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
peewee=3.16.0=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=pypi_0
pip=22.3.1=py310hca03da5_0
platformdirs=3.5.1=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.1=pypi_0
psutil=5.9.0=py310h1a28f6b_0
psycopg2=2.9.5=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pymysql=1.0.2=pypi_0
python=3.10.0=hbdb9e5c_5
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-telegram-bot=20.1=pypi_0
pyzmq=25.0.2=py310h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=65.6.3=py310hca03da5_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
sqlite=3.40.1=h7a7dc30_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torch=2.0.1=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.2=py310h1a28f6b_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.6.1=hd8ed1ab_0
typing_extensions=4.6.1=pyha770c72_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.3.4=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h80987f9_1
zeromq=4.3.4=hbdafb3b_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5a0b063_0
